fixed some warnings in the MATLAB bindings
git-svn-id: https://libgdf.svn.sourceforge.net/svnroot/libgdf/trunk@24 00a3b082-7228-46c3-a6fa-0e10d54cf57f

+++ matlab/matlab_tools/Commands.h
@@ -25,7 +25,7 @@ class Command
 public:
     virtual void execute( mxArray *plhs[], const mxArray *prhs[] ) = 0;
 
-    void operator()( int nlhs, mxArray *plhs[], int nrhs, const mxArray *prhs[] )
+    void operator()( size_t nlhs, mxArray *plhs[], size_t nrhs, const mxArray *prhs[] )
     {
         if( m_nlhs != nlhs )
             throw std::invalid_argument( "Invalid number of output arguments." );
